requirements.txt
# OpenEnv-Orbital-Command — requirements
gymnasium>=0.29.0
pydantic>=2.0
gradio>=4.20.0
numpy>=1.24.0
plotly>=5.0.0          # reward chart (optional but recommended)
python-dotenv>=1.0.0   # .env file support

# LLM backends — install only what you use
google-genai>=0.5.0            # Google Gemini (new SDK — recommended)
# google-generativeai>=0.7.0   # Google Gemini (legacy SDK, also works)
huggingface_hub>=0.21.0
